requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
addict==2.4.0
    # via open3d
ansi2html==1.8.0
    # via dash
asttokens==2.2.1
    # via stack-data
attrs==23.1.0
    # via
    #   fiona
    #   jsonschema
    #   referencing
backcall==0.2.0
    # via ipython
certifi==2023.7.22
    # via
    #   fiona
    #   pyproj
    #   requests
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   flask
    #   whitebox
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
comm==0.1.3
    # via ipykernel
configargparse==1.7
    # via open3d
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
dash==2.11.1
    # via open3d
dash-core-components==2.0.0
    # via dash
dash-html-components==2.0.0
    # via dash
dash-table==5.0.0
    # via dash
debugpy==1.6.7
    # via ipykernel
decorator==5.1.1
    # via ipython
executing==1.2.0
    # via stack-data
fastjsonschema==2.18.0
    # via nbformat
fiona==1.9.4.post1
    # via geopandas
flask==2.2.5
    # via dash
fonttools==4.41.1
    # via matplotlib
geopandas==0.13.2
    # via -r requirements.in
idna==3.4
    # via requests
importlib-metadata==6.8.0
    # via
    #   fiona
    #   flask
    #   jupyter-client
importlib-resources==6.0.0
    # via
    #   jsonschema
    #   jsonschema-specifications
    #   matplotlib
ipykernel==6.24.0
    # via ipywidgets
ipython==8.12.2
    # via
    #   ipykernel
    #   ipywidgets
ipywidgets==8.0.7
    # via open3d
itsdangerous==2.1.2
    # via flask
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via flask
joblib==1.3.1
    # via scikit-learn
jsonschema==4.18.4
    # via nbformat
jsonschema-specifications==2023.7.1
    # via jsonschema
jupyter-client==8.3.0
    # via ipykernel
jupyter-core==5.3.1
    # via
    #   ipykernel
    #   jupyter-client
    #   nbformat
jupyterlab-widgets==3.0.8
    # via ipywidgets
kiwisolver==1.4.4
    # via matplotlib
laspy==2.5.0
    # via -r requirements.in
loguru==0.7.0
    # via -r requirements.in
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.7.2
    # via
    #   -r requirements.in
    #   open3d
matplotlib-inline==0.1.6
    # via
    #   ipykernel
    #   ipython
nbformat==5.7.0
    # via open3d
nest-asyncio==1.5.6
    # via
    #   dash
    #   ipykernel
numpy==1.24.4
    # via
    #   -r requirements.in
    #   contourpy
    #   laspy
    #   matplotlib
    #   open3d
    #   pandas
    #   pyquaternion
    #   scikit-learn
    #   scipy
    #   shapely
open3d==0.17.0
    # via -r requirements.in
packaging==23.1
    # via
    #   geopandas
    #   ipykernel
    #   matplotlib
    #   plotly
pandas==2.0.3
    # via
    #   geopandas
    #   open3d
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.0.0
    # via
    #   matplotlib
    #   open3d
pkgutil-resolve-name==1.3.10
    # via jsonschema
platformdirs==3.9.1
    # via jupyter-core
plotly==5.15.0
    # via dash
prompt-toolkit==3.0.39
    # via ipython
psutil==5.9.5
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.15.1
    # via ipython
pyparsing==3.0.9
    # via matplotlib
pyproj==3.5.0
    # via geopandas
pyquaternion==0.9.9
    # via open3d
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   open3d
pyzmq==25.1.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.30.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements.in
    #   dash
retrying==1.3.4
    # via dash
rpds-py==0.9.2
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.3.0
    # via open3d
scipy==1.10.1
    # via scikit-learn
shapely==2.0.1
    # via
    #   -r requirements.in
    #   geopandas
six==1.16.0
    # via
    #   asttokens
    #   fiona
    #   python-dateutil
    #   retrying
stack-data==0.6.2
    # via ipython
tenacity==8.2.2
    # via plotly
threadpoolctl==3.2.0
    # via scikit-learn
tornado==6.3.2
    # via
    #   ipykernel
    #   jupyter-client
tqdm==4.65.0
    # via open3d
traitlets==5.9.0
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   nbformat
typing-extensions==4.7.1
    # via
    #   dash
    #   ipython
tzdata==2023.3
    # via pandas
urllib3==2.0.4
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
werkzeug==2.2.3
    # via
    #   dash
    #   flask
    #   open3d
whitebox==2.3.1
    # via -r requirements.in
widgetsnbextension==4.0.8
    # via ipywidgets
zipp==3.16.2
    # via
    #   importlib-metadata
    #   importlib-resources
